chore(tokenize): extract to tokenize
